Adapt the code to work perfectly with the library used

+++ app/src/main/java/ch/epfl/sweng/opengm/polls/results/HoloGraphLibrary/Bar.java
@@ -114,7 +114,7 @@ public class Bar {
         if (mValueString != null) {
             return mValueString;
         } else {
-            return String.valueOf(mValue);
+            return String.valueOf((int) mValue);
         }
     }
